Prepare ALO 0.14.3

+++ Sources/ALO/DJLiveDeckView.swift
@@ -71,16 +71,16 @@ struct DJLiveDeckView: View {
             Canvas { context, size in
                 let peaks = live.waveform
                 guard !peaks.isEmpty else { return }
-                let step = size.width / Double(peaks.count)
+                let step = size.width / CGFloat(peaks.count)
                 var bars = Path()
                 for (index, peak) in peaks.enumerated() {
-                    let height = max(1, Double(peak) * size.height * 0.9)
-                    bars.addRoundedRect(in: CGRect(x: Double(index) * step, y: (size.height - height) / 2,
+                    let height = max(1, CGFloat(peak) * size.height * 0.9)
+                    bars.addRoundedRect(in: CGRect(x: CGFloat(index) * step, y: (size.height - height) / 2,
                                                    width: max(1, step - 1), height: height), cornerSize: CGSize(width: 1, height: 1))
                 }
                 context.fill(bars, with: .color(color.opacity(0.7)))
                 let delay = scrubbing ? scrubDelay : live.delaySeconds
-                let x = size.width * (1 - min(1, delay / max(0.001, live.historyDuration)))
+                let x = size.width * CGFloat(1 - min(1, delay / max(0.001, live.historyDuration)))
                 var playhead = Path()
                 playhead.move(to: CGPoint(x: x, y: 0)); playhead.addLine(to: CGPoint(x: x, y: size.height))
                 context.stroke(playhead, with: .color(.white), lineWidth: 2)
